replace custom function with in built String()
